desktop_test/requirements.txt
pytest==7.4.4
pyautogui==0.9.54
PyQt6==6.6.1
PySide6==6.6.1
Pillow==11.1.0
pytest-html==4.1.1
opencv-python==4.11.0.86
numpy==1.26.3
loguru==0.7.2
python-magic==0.4.27 